pyproject.toml
[project]
name = "lexora-link"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.0",
    "pydantic-ai>=1.63.0",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "chromadb>=0.6.0",
    "ruff>=0.15.2",
    "structlog>=25.5.0",
    "trafilatura>=2.0.0",
    "uvicorn>=0.41.0",
    "google-genai>=1.64.0",
    "pyyaml>=6.0.3",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "aiofiles>=25.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests/unit"]
markers = ["evals: requires a live LLM API key; run with pytest tests/evals/"]

[dependency-groups]
dev = [
    "pydantic-ai[evals]>=1.63.0",
]
